+++ native/index.js
@@ -671,6 +671,8 @@ if (!nativeBinding) {
 }
 
 module.exports = nativeBinding
+module.exports.cloneFileExclusive = nativeBinding.cloneFileExclusive
+module.exports.copyFileRangeExclusive = nativeBinding.copyFileRangeExclusive
 module.exports.extractArchiveNative = nativeBinding.extractArchiveNative
 module.exports.fstatIdentity = nativeBinding.fstatIdentity
 module.exports.inspectArchiveNative = nativeBinding.inspectArchiveNative
@@ -679,3 +681,4 @@ module.exports.mkdirBeneath = nativeBinding.mkdirBeneath
 module.exports.openBeneath = nativeBinding.openBeneath
 module.exports.readArchiveEntryNative = nativeBinding.readArchiveEntryNative
 module.exports.renameNoReplace = nativeBinding.renameNoReplace
+module.exports.sha256File = nativeBinding.sha256File